user data fetching all moves to App.jsx thanks to useEffet. Need to add listener still though
