+++ src/App.jsx
@@ -9,7 +9,9 @@ import { useEffect, useState } from 'react'
 //import './App.css'
 
 // Import created modules
-import { redirectToAuthCodeFlow } from './modules/spotify_connect'
+import { redirectToAuthCodeFlow, getAccessToken } from './modules/spotify_connect'
+import { fetchAllSongs, fetchProfile, fetchTopTracks, filterUnplayables } from './modules/spotify_fetch.js'
+
 
 const Connecter = (props) => {
 
@@ -110,33 +112,81 @@ const App = (props) => {
   // console.log('App props: ', props)
 
   const appClientId = 'e61711cbd130408abf2d471288b77e87';
-  const redirectUrl = 'http://localhost:5173/'
+  const redirectUrl = 'http://localhost:5174/'
+  // .env variables should defined as this but it doesn't work for now.
   // const appClientId = env.spotifyAppClientId;
   // const redirectUrl = env.redirectUrl; // IMPORTANT: IT HAS TO BE A REDIRECT URI ON THE APP SETTINGS ONLINE
+  
+  // Actually we could get rid of the access token code here if we store the data and don't fetch it twice...
+  const accessToken = localStorage.getItem("access_token");
+  // console.log("Access token is stored at: " + accessToken);
+
+  // Retrieve the user code from the URL just after the Spotify opt-in.
+  const params = new URLSearchParams(window.location.search)
+  const URLparamCode = params.get('code');
+  const storedUsedCode = localStorage.getItem('spotify_user_code')
+  // WARNING: THE USER CODE REMAINS IN THE URL. SHOULD PROBABLY REDIRECT TO PAGE WITHOUT IT, BUT... WHEN I'LL HAVE A DATABASE
+  const isCodeUsed = URLparamCode === storedUsedCode;
+
+  // I had to define all userData properties since they are used in App.jsx. 
+  // There's probably a better way, either through class, or JSON loading
+  // Since it's the data per user, it should be stored in a JSON server, encapsulated for each user, and not local storaged
+  let initUserData = localStorage.getItem('user_data')?
+    JSON.parse(localStorage.getItem('user_data'))
+    : { 
+      fetched: false,
+      profile: {displayName: '', email: '', country:'', external_urls: {spotify: ''}, images: []}, 
+      topTracks: [], 
+      allSongs: [], 
+      unplayables: [],
+    }
 
   const connectToSpotify = () => {
     redirectToAuthCodeFlow(appClientId, redirectUrl)
   }
 
   // DECIDED TO LOAD THE USERDATA DIRECTLY FROM MAIN
-  // console.log('stored user_data:', localStorage.getItem('user_data'));
-  // const [userData, setUserData] = useState({props.userData})
-  // setUserData(JSON.parse(localStorage.getItem('user_data')))
-  // THIS DOESN'T WORK BECAUSE IT'S UNABLE TO GET THE SINGLE VARIABLES EVERYTIME...
+  const [userData, setUserData] = useState(initUserData)
+  // Maybe I'll have to use useEffect if the userData is not defined yet...
   // useEffect(() => {
-  //   if (localStorage.getItem('user_data')) {
-  //     setUserData(JSON.parse(localStorage.getItem('user_data')))
-  //   }
+  //   setUserData(userData)
   // }, [])
 
-  // let userData = { profile: null, topTracks:null, allSongs:null, unplayables:null }
-  // if (localStorage.getItem('user_data')) {
-  //   const userData = JSON.parse(localStorage.getItem('user_data'))
-  // } else {
-  //   const userData = { profile: null, topTracks:null, allSongs:null, unplayables:null }
-  // }
-  // const userData = JSON.parse(localStorage.getItem('user_data'));
-  console.log('in App userData:', props.userData);
+
+  // useEffect to run the code only once
+  useEffect(() => {
+    // Théoriquement, on aurait même pas besoin de faire un localStorage, il faut juste passer la variable à l'app
+    // Ici on utilise userData.fetched mais il faudrait plutôt un évènement de refresher
+    if (!userData.fetched && URLparamCode && !isCodeUsed) {
+      // REMOVED CONDITION (!accessToken || accessToken === 'undefined')
+      // since we use the accessToken only once - for now
+      console.log('Getting access token...')
+      getAccessToken(appClientId, URLparamCode, redirectUrl).then(accessToken => {
+        const profile = fetchProfile(accessToken)
+        const topTracks = fetchTopTracks(accessToken)
+        const allSongs = fetchAllSongs(accessToken)
+        const unplayables = allSongs
+          .then(songs => { 
+            const filteredUnplayables = filterUnplayables(songs);
+            return filteredUnplayables;
+              // console.log('Post then unplayables:', filterUnplayables(songs))
+            })
+          .catch(error => console.error('Failed to fetch all tracks', error))
+    
+        console.log('Post then unplayables, probably still waiting:', unplayables)
+    
+        Promise.all([profile, topTracks, allSongs, unplayables]).then(([profile, topTracks, allSongs, unplayables]) => {
+          console.log('Starting promise to pass all user data')
+          setUserData({ fetched: true, profile, topTracks, allSongs, unplayables })
+          console.log('After Promise userData:', userData);
+          localStorage.setItem('user_data', JSON.stringify(userData));
+        })
+      })
+      // Can't reuse the same user code, storing it local storage to avoid confusion
+      localStorage.setItem('spotify_user_code', URLparamCode);
+    } 
+  })
+
 
   return (
     <>
@@ -144,9 +194,9 @@ const App = (props) => {
       <h1>Display your Spotify profile data</h1>
 
       <Connecter onClick={connectToSpotify}/> {/* Find a a way to hide if existing userData or change to REFRESH DATA */}
-      <Profile profile={props.userData.profile}/>
-      <TopTracks topTracks={props.userData.topTracks}/>
-      <UnplayableTracks unplayables={props.userData.unplayables}/>
+      <Profile profile={userData.profile}/>
+      <TopTracks topTracks={userData.topTracks}/>
+      <UnplayableTracks unplayables={userData.unplayables}/>
 
       <br/>
     </>
